+++ features/support/env.rb
@@ -1,4 +1,5 @@
 require 'aruba/cucumber'
+require 'pry'
 
 Before do
   @aruba_timeout_seconds = 60 * 2